Refactor models into per-model files
