+++ models/__init__.py
@@ -1,4 +1,10 @@
 # -*- coding: utf-8 -*-
 
-from . import models
-from . import users
+from . import erasmus_persona
+from . import erasmus_ciclo
+from . import erasmus_movilidad
+from . import erasmus_codigo
+from . import erasmus_pais
+from . import res_partner
+from . import res_users
+from . import res_country
