+++ models/erasmus_persona.py
@@ -3,13 +3,6 @@ from odoo import models, fields, api
 from odoo.exceptions import ValidationError
 import logging
 import unicodedata
-import mimetypes
-
-
-class ResPartner(models.Model):
-    _inherit = 'res.partner'
-
-    erasmus_persona_id = fields.Many2one('erasmus.persona', string='Persona Erasmus Vinculada', readonly=True)
 
 
 class ErasmusPersona(models.Model):
@@ -1287,525 +1280,3 @@ class ErasmusPersona(models.Model):
                 rec.ciclo_formativo_id = False
                 rec.ciclo_formativo = False
 
-
-class ErasmusCiclo(models.Model):
-    _name = 'erasmus.ciclo'
-    _description = 'Ciclo Formativo Erasmus'
-    _order = 'familia_profesional, name'
-
-    name = fields.Char(required=True)
-    code = fields.Char(required=True, index=True)
-    familia_profesional = fields.Selection([
-        ('informatica', 'Informática y Comunicaciones'),
-        ('administracion', 'Administración y Gestión'),
-        ('comercio', 'Comercio y Marketing'),
-        ('transporte', 'Transporte y Mantenimiento de Vehículos'),
-        ('sanidad', 'Sanidad'),
-        ('servicios', 'Servicios Socioculturales y a la Comunidad'),
-        # ('otros', 'Otros / Formación complementaria'),
-    ], required=True, index=True)
-
-    # Nivel del ciclo (para filtrar por nivel de impartición)
-    nivel = fields.Selection([
-        ('fpb', 'FP Básica'),
-        ('cfgm', 'Grado Medio'),
-        ('cfgs', 'Grado Superior'),
-        ('egm', 'Especialización Grado Medio'),
-        ('egs', 'Especialización Grado Superior'),
-    ], index=True)
-
-    _sql_constraints = [
-        ('uniq_code', 'unique(code)', 'El código del ciclo debe ser único.')
-    ]
-
-
-class ErasmusMovilidad(models.Model):
-    _name = 'erasmus.movilidad'
-    _description = 'Movilidad Erasmus'
-    _order = 'id desc'
-
-    # Enlace y tipo
-    persona_id = fields.Many2one('erasmus.persona', string='Contacto', required=True, ondelete='cascade', index=True)
-    tipo_interno = fields.Selection([
-        ('estudiante', 'Estudiante'),
-        ('profesor', 'Profesor'),
-        ('acompaniante', 'Acompañante')
-    ], string='Tipo interno de estudiante', default=lambda self: self.env.context.get('default_tipo_interno'), required=True, index=True)
-
-    # Datos principales
-    curso_academico = fields.Char(string='Curso académico')
-    cuenta_bancaria = fields.Char(string='Cuenta bancaria')
-    num_adjuntos_dni = fields.Selection([
-        ('1', '1 adjunto'),
-        ('2', '2 adjuntos'),
-    ], string='Número adjuntos DNI', default='1', required=True, help='Si eliges 1 adjunto, sube ambas caras en un único archivo. Si eliges 2 adjuntos, sube anverso y reverso por separado.')
-    dni = fields.Binary(string='DNI (anverso o archivo único)', attachment=True)
-    dni_filename = fields.Char(string='Nombre archivo DNI')
-    dni2 = fields.Binary(string='DNI 2', attachment=True)
-    dni2_filename = fields.Char(string='Nombre archivo DNI 2')
-    cert_titularidad_bancaria = fields.Binary(string='Certificado de titularidad bancaria', attachment=True)
-    cert_titularidad_bancaria_filename = fields.Char(string='Nombre archivo Certificado titularidad')
-    tipo_movilidad = fields.Char(string='Tipo movilidad')
-    es_formacion = fields.Boolean(string='Es formación')
-
-    # Datos empresa / viaje
-    fecha_ida = fields.Date(string='Fecha de ida')
-    fecha_vuelta = fields.Date(string='Fecha de vuelta')
-    duracion_actividad_dias = fields.Integer(string='Duración actividad (días)', compute='_compute_duracion', store=True)
-    que_dias = fields.Char(string='Qué días')
-    pais_destino = fields.Many2one('res.country', string='País de destino')
-    nombre_empresa = fields.Char(string='Nombre empresa')
-    direccion_empresa = fields.Char(string='Dirección empresa')
-    ciudad_empresa = fields.Char(string='Ciudad empresa')
-    persona_contacto_empresa = fields.Char(string='Nombre y apellido de la persona de contacto')
-    cargo_empresa = fields.Char(string='Cargo en la empresa')
-    email_empresa = fields.Char(string='Correo-e empresa')
-    telefono_empresa = fields.Char(string='Teléfono empresa')
-
-    # Datos profesor / acompañante
-    obj_generales = fields.Text(string='Objetivos generales de la movilidad')
-    actividades_realizar = fields.Text(string='Actividades a realizar')
-    resultados_impacto = fields.Text(string='Resultados e impacto previstos')
-    informacion_extra_prof = fields.Text(string='Información extra')
-    digital_skill = fields.Char(string='Digital skill')
-    medio_transporte = fields.Char(string='Medio de transporte')
-
-    # Datos estudiantes (nueva lógica de países Erasmus)
-    # Campos antiguos conservados por compatibilidad (invisibles en vistas)
-    pref_destino_1 = fields.Many2one('res.country', string='Preferencia país de destino 1', help='[LEGACY] No usar, sustituido por pref_pais_1_id')
-    pref_destino_2 = fields.Many2one('res.country', string='Preferencia país de destino 2', help='[LEGACY] No usar, sustituido por pref_pais_2_id')
-    pref_destino_3 = fields.Many2one('res.country', string='Preferencia país de destino 3', help='[LEGACY] No usar, sustituido por pref_pais_3_id')
-    # Almacenamiento real en res.country (requerimiento)
-    pref_pais_1_id = fields.Many2one('res.country', string='Preferencia país 1')
-    pref_pais_2_id = fields.Many2one('res.country', string='Preferencia país 2')
-    pref_pais_3_id = fields.Many2one('res.country', string='Preferencia país 3')
-    # Ayuda para dominios cuando el selector UI usa erasmus.pais
-    allowed_country_ids = fields.Many2many('res.country', string='Países permitidos', compute='_compute_allowed_countries', store=False)
-    allowed_pais_ids = fields.Many2many('erasmus.pais', string='Países permitidos (UI)', compute='_compute_allowed_paises', store=False)
-    # Campos UI (no almacenan, abren 'Search More' sobre erasmus.pais y se sincronizan con pref_pais_* res.country)
-    ui_pref_pais_1_id = fields.Many2one('erasmus.pais', string='Preferencia país 1', compute='_compute_ui_pref_pais_1', inverse='_inverse_ui_pref_pais_1', store=False)
-    ui_pref_pais_2_id = fields.Many2one('erasmus.pais', string='Preferencia país 2', compute='_compute_ui_pref_pais_2', inverse='_inverse_ui_pref_pais_2', store=False)
-    ui_pref_pais_3_id = fields.Many2one('erasmus.pais', string='Preferencia país 3', compute='_compute_ui_pref_pais_3', inverse='_inverse_ui_pref_pais_3', store=False)
-    informacion_extra_est = fields.Text(string='Información extra/peticiones')
-    alumno_menos_oportunidad = fields.Boolean(string='¿Alumno con menos oportunidad?')
-    motivo_menos_oportunidad = fields.Text(string='Motivo menor oportunidad')
-    contacto_origen = fields.Char(string='Contacto en Origen')
-    parentesco = fields.Char(string='Parentesco')
-    telefono_contacto = fields.Char(string='Teléfono del Contacto')
-    recien_titulado = fields.Boolean(string='Recién titulado')
-    curriculum_ingles = fields.Binary(string='Curriculum en inglés', attachment=True)
-    curriculum_ingles_filename = fields.Char(string='Nombre archivo CV inglés')
-    carta_presentacion_ingles = fields.Binary(string='Carta de presentación en inglés', attachment=True)
-    carta_presentacion_ingles_filename = fields.Char(string='Nombre archivo Carta presentación')
-    certificado_1 = fields.Binary(string='Certificado 1', attachment=True)
-    certificado_1_filename = fields.Char(string='Nombre archivo Certificado 1')
-    certificado_2 = fields.Binary(string='Certificado 2', attachment=True)
-    certificado_2_filename = fields.Char(string='Nombre archivo Certificado 2')
-    certificado_3 = fields.Binary(string='Certificado 3', attachment=True)
-    certificado_3_filename = fields.Char(string='Nombre archivo Certificado 3')
-
-    # Estado
-    estado_datos = fields.Selection([
-        ('borrador', 'Borrador'),
-        ('pendiente', 'Pendiente'),
-        ('completo', 'Completo'),
-    ], string='Estado datos', default='borrador', index=True)
-
-    @api.depends('fecha_ida', 'fecha_vuelta')
-    def _compute_duracion(self):
-        for rec in self:
-            if rec.fecha_ida and rec.fecha_vuelta and rec.fecha_vuelta >= rec.fecha_ida:
-                rec.duracion_actividad_dias = (rec.fecha_vuelta - rec.fecha_ida).days
-            else:
-                rec.duracion_actividad_dias = 0
-
-    @api.onchange('num_adjuntos_dni')
-    def _onchange_num_adjuntos_dni(self):
-        for rec in self:
-            # Si no son 2 adjuntos, limpiamos el campo del reverso
-            if rec.num_adjuntos_dni != '2':
-                rec.dni2 = False
-
-    @api.onchange('persona_id')
-    def _onchange_persona_id(self):
-        for rec in self:
-            if rec.persona_id:
-                # Sincroniza automáticamente el tipo con el de la persona
-                rec.tipo_interno = rec.persona_id.tipo_interno or rec.tipo_interno
-
-    def _get_country_scope_sets(self):
-        Pais = self.env['erasmus.pais'].sudo()
-        ambos_ids = Pais.search([('selection_scope', '=', 'ambos'), ('active', '=', True)]).mapped('country_id').ids
-        estu_ids = Pais.search([('selection_scope', '=', 'estudiante'), ('active', '=', True)]).mapped('country_id').ids
-        return set(ambos_ids), set(estu_ids)
-
-    @api.depends('pref_pais_1_id', 'pref_pais_2_id', 'pref_pais_3_id')
-    def _compute_allowed_countries(self):
-        ambos_set, estu_set = self._get_country_scope_sets()
-        for rec in self:
-            selected_list = [rec.pref_pais_1_id, rec.pref_pais_2_id, rec.pref_pais_3_id]
-            selected = {c.id for c in selected_list if c}
-            all_three_set = all(selected_list)
-            has_student_only = any(cid in estu_set for cid in selected)
-            if has_student_only or not all_three_set:
-                allowed = ambos_set | estu_set
-            else:
-                allowed = ambos_set
-            rec.allowed_country_ids = [(6, 0, list(allowed))]
-
-    @api.depends('pref_pais_1_id', 'pref_pais_2_id', 'pref_pais_3_id')
-    def _compute_allowed_paises(self):
-        """Igual que allowed_country_ids, pero devolviendo ids de erasmus.pais para usar en los campos UI."""
-        Pais = self.env['erasmus.pais'].sudo()
-        ambos_set_c, estu_set_c = self._get_country_scope_sets()
-        # Mapear countries permitidos -> paises (erasmus.pais)
-        ambos_pais = set(Pais.search([('country_id', 'in', list(ambos_set_c)), ('active', '=', True)]).ids)
-        estu_pais = set(Pais.search([('country_id', 'in', list(estu_set_c)), ('active', '=', True)]).ids)
-        for rec in self:
-            selected_list = [rec.pref_pais_1_id, rec.pref_pais_2_id, rec.pref_pais_3_id]
-            selected = {c.id for c in selected_list if c}
-            all_three_set = all(selected_list)
-            has_student_only = any(cid in estu_set_c for cid in selected)
-            if has_student_only or not all_three_set:
-                allowed = ambos_pais | estu_pais
-            else:
-                allowed = ambos_pais
-            rec.allowed_pais_ids = [(6, 0, list(allowed))]
-
-    @api.onchange('pref_pais_1_id', 'pref_pais_2_id', 'pref_pais_3_id')
-    def _onchange_pref_paises(self):
-        ambos_set, estu_set = self._get_country_scope_sets()
-        for rec in self:
-            selected_list = [rec.pref_pais_1_id, rec.pref_pais_2_id, rec.pref_pais_3_id]
-            selected = {c.id for c in selected_list if c}
-            all_three_set = all(selected_list)
-            has_student_only = any(cid in estu_set for cid in selected)
-            if has_student_only or not all_three_set:
-                allowed = list(ambos_set | estu_set)
-            else:
-                allowed = list(ambos_set)
-            domain = [('id', 'in', allowed)]
-            return {'domain': {'pref_pais_1_id': domain, 'pref_pais_2_id': domain, 'pref_pais_3_id': domain}}
-
-    # Cálculo y sincronización de campos UI (erasmus.pais <-> res.country)
-    def _map_country_to_pais(self, country):
-        if not country:
-            return False
-        return self.env['erasmus.pais'].sudo().search([('country_id', '=', country.id), ('active', '=', True)], limit=1)
-
-    @api.depends('pref_pais_1_id')
-    def _compute_ui_pref_pais_1(self):
-        for rec in self:
-            rec.ui_pref_pais_1_id = rec._map_country_to_pais(rec.pref_pais_1_id)
-
-    @api.depends('pref_pais_2_id')
-    def _compute_ui_pref_pais_2(self):
-        for rec in self:
-            rec.ui_pref_pais_2_id = rec._map_country_to_pais(rec.pref_pais_2_id)
-
-    @api.depends('pref_pais_3_id')
-    def _compute_ui_pref_pais_3(self):
-        for rec in self:
-            rec.ui_pref_pais_3_id = rec._map_country_to_pais(rec.pref_pais_3_id)
-
-    def _inverse_ui_pref_pais_1(self):
-        for rec in self:
-            rec.pref_pais_1_id = rec.ui_pref_pais_1_id.country_id if rec.ui_pref_pais_1_id else False
-
-    def _inverse_ui_pref_pais_2(self):
-        for rec in self:
-            rec.pref_pais_2_id = rec.ui_pref_pais_2_id.country_id if rec.ui_pref_pais_2_id else False
-
-    def _inverse_ui_pref_pais_3(self):
-        for rec in self:
-            rec.pref_pais_3_id = rec.ui_pref_pais_3_id.country_id if rec.ui_pref_pais_3_id else False
-
-    @api.onchange('ui_pref_pais_1_id')
-    def _onchange_ui_pref_pais_1(self):
-        for rec in self:
-            rec.pref_pais_1_id = rec.ui_pref_pais_1_id.country_id if rec.ui_pref_pais_1_id else False
-
-    @api.onchange('ui_pref_pais_2_id')
-    def _onchange_ui_pref_pais_2(self):
-        for rec in self:
-            rec.pref_pais_2_id = rec.ui_pref_pais_2_id.country_id if rec.ui_pref_pais_2_id else False
-
-    @api.onchange('ui_pref_pais_3_id')
-    def _onchange_ui_pref_pais_3(self):
-        for rec in self:
-            rec.pref_pais_3_id = rec.ui_pref_pais_3_id.country_id if rec.ui_pref_pais_3_id else False
-
-    @api.model_create_multi
-    def create(self, vals_list):
-        # Asegurar tipo_interno desde contexto o persona
-        for vals in vals_list:
-            if not vals.get('tipo_interno'):
-                pid = vals.get('persona_id') or self.env.context.get('default_persona_id')
-                if pid:
-                    persona = self.env['erasmus.persona'].browse(pid)
-                    vals['tipo_interno'] = persona.tipo_interno or self.env.context.get('default_tipo_interno') or 'estudiante'
-                else:
-                    vals['tipo_interno'] = self.env.context.get('default_tipo_interno') or 'estudiante'
-        return super().create(vals_list)
-
-    def write(self, vals):
-        # Si cambia la persona y no se indicó tipo, sincronizar con el tipo de la nueva persona
-        if 'persona_id' in vals and 'tipo_interno' not in vals and vals.get('persona_id'):
-            persona = self.env['erasmus.persona'].browse(vals['persona_id'])
-            vals = dict(vals, tipo_interno=persona.tipo_interno or 'estudiante')
-        return super().write(vals)
-
-    def action_download_binary(self):
-        self.ensure_one()
-        field = self.env.context.get('binary_field')
-        if not field:
-            raise ValueError('No se indicó el campo binario a descargar (binary_field).')
-        value = getattr(self, field, False)
-        if not value:
-            raise ValueError('No hay archivo cargado para descargar.')
-        # Determinar nombre de archivo preferido
-        filename_field = f"{field}_filename"
-        filename = getattr(self, filename_field, None)
-        if not filename:
-            # Intentar recuperar el adjunto para obtener 'name' o inferir extensión desde mimetype
-            att = self.env['ir.attachment'].search([
-                ('res_model', '=', self._name),
-                ('res_id', '=', self.id),
-                ('res_field', '=', field),
-            ], order='id desc', limit=1)
-            if att:
-                filename = att.name or None
-                # Si no tiene extensión pero hay mimetype, intentar adivinar
-                if filename and '.' not in filename and att.mimetype:
-                    ext = mimetypes.guess_extension(att.mimetype) or ''
-                    # Corrección común: image/jpeg suele devolver .jpe
-                    if ext == '.jpe':
-                        ext = '.jpg'
-                    filename = f"{filename}{ext}"
-        # Fallback final
-        if not filename:
-            # Intentar una extensión genérica a partir del contenido si existe mimetype en attachments previos
-            filename = f"{field}.bin"
-        url = f"/web/content/{self._name}/{self.id}/{field}/{filename}?download=true"
-        return {
-            'type': 'ir.actions.act_url',
-            'url': url,
-            'target': 'self',
-        }
-
-
-class ErasmusCodigo(models.Model):
-    _name = 'erasmus.codigo'
-    _description = 'Catálogo de Códigos (Programa, Código Erasmus, ISCED-F)'
-    _order = 'key, ciclo_id, id'
-
-    key = fields.Selection([
-        ('programa', 'Programa'),
-        ('codigo_erasmus', 'Código Erasmus'),
-        ('codigo_iscedf', 'Código ISCED-F'),
-    ], required=True, index=True)
-    ciclo_id = fields.Many2one('erasmus.ciclo', string='Ciclo', index=True, help='Opcional: específica por ciclo. Si está vacío, actúa como valor global por defecto.')
-    valor = fields.Char(string='Valor', required=True)
-    active = fields.Boolean(default=True)
-    note = fields.Char(string='Descripción', help='Notas u observaciones del código')
-
-    _sql_constraints = [
-        ('uniq_key_ciclo', 'unique(key, ciclo_id)', 'Ya existe un código para esa clave y ciclo.'),
-    ]
-
- 
-class ResPartner(models.Model):
-    _inherit = 'res.partner'
-
-    def write(self, vals):
-        # PRE: Cascada de archivado hacia persona/usuario (usuario primero) si cambia active en contacto
-        if 'active' in vals and not self.env.context.get('skip_partner_active_cascade'):
-            new_active = bool(vals.get('active'))
-            personas = self.env['erasmus.persona'].with_context(active_test=False).sudo().search([('partner_id', 'in', self.ids)])
-            if personas:
-                try:
-                    # Evitar que persona.write reescriba el partner durante este write
-                    personas.sudo().with_context(skip_partner_back_write=True).write({'active': new_active})
-                except Exception:
-                    pass
-        res = super().write(vals)
-        # Sincronizar cambios a persona salvo que venga con flag para evitar bucles
-        if not self.env.context.get('skip_persona_sync'):
-            # Considerar más claves: name, vat, email, mobile, y dirección
-            keys = {'name', 'vat', 'email', 'mobile', 'street', 'street2', 'zip', 'city', 'state_id', 'country_id'}
-            if any(k in vals for k in keys):
-                personas = self.env['erasmus.persona'].search([('partner_id', 'in', self.ids)])
-                for persona in personas:
-                    updates = {}
-                    if 'name' in vals and vals.get('name'):
-                        parts = (vals['name'] or '').split()
-                        if parts:
-                            updates['nombre'] = parts[0]
-                            updates['apellido1'] = parts[-1] if len(parts) > 1 else False
-                            updates['apellido2'] = ' '.join(parts[1:-1]) if len(parts) > 2 else False
-                    if 'vat' in vals:
-                        updates['nif'] = vals.get('vat') or False
-                    if 'email' in vals:
-                        updates['email'] = vals.get('email') or False
-                    if 'mobile' in vals:
-                        updates['movil'] = vals.get('mobile') or False
-                    for k in ('street', 'street2', 'zip', 'city', 'state_id', 'country_id'):
-                        if k in vals:
-                            updates[k] = vals.get(k) or False
-                    if updates:
-                        # from_partner: evitar que persona.write vuelva a escribir partner
-                        # skip_user_sync=False para que actualice usuario si aplica (email/nombre)
-                        persona.with_context(skip_partner_sync=True, from_partner=True).write(updates)
-        return res
-
-    @api.model_create_multi
-    def create(self, vals_list):
-        records = super().create(vals_list)
-        # Evitar auto-creación si viene desde erasmus.persona
-        if self.env.context.get('skip_auto_persona'):
-            return records
-        Persona = self.env['erasmus.persona']
-        for partner in records:
-            try:
-                # No duplicar si ya hay vinculada
-                existing = Persona.search([('partner_id', '=', partner.id)], limit=1)
-                if existing:
-                    partner.erasmus_persona_id = existing.id
-                    continue
-                name = (partner.name or '').strip()
-                parts = name.split()
-                # Crear persona mínima y evitar escribir campos relacionados (email/móvil/dirección)
-                # para no disparar partner.write durante partner.create
-                persona = Persona.sudo().create({
-                    'tipo_interno': 'no_asignado',
-                    'partner_id': partner.id,
-                    'nombre': parts[0] if parts else False,
-                    'apellido1': parts[-1] if len(parts) > 1 else False,
-                    'apellido2': ' '.join(parts[1:-1]) if len(parts) > 2 else False,
-                    # nif/email/móvil/dirección se reflejarán vía campos related sin escribir de nuevo el partner
-                })
-                partner.erasmus_persona_id = persona.id
-            except Exception:
-                # No bloquear creación de partner por errores no críticos
-                continue
-        return records
-
-    def action_open_or_create_persona(self):
-        self.ensure_one()
-        Persona = self.env['erasmus.persona'].with_context(active_test=False)
-        persona = Persona.search([('partner_id', '=', self.id)], limit=1)
-        if not persona:
-            name = (self.name or '').strip()
-            parts = name.split()
-            persona = Persona.sudo().create({
-                'tipo_interno': 'no_asignado',
-                'partner_id': self.id,
-                'nombre': parts[0] if parts else False,
-                'apellido1': parts[-1] if len(parts) > 1 else False,
-                'apellido2': ' '.join(parts[1:-1]) if len(parts) > 2 else False,
-            })
-            self.erasmus_persona_id = persona.id
-        return {
-            'type': 'ir.actions.act_window',
-            'name': 'Persona Erasmus',
-            'res_model': 'erasmus.persona',
-            'view_mode': 'form',
-            'res_id': persona.id,
-            'target': 'current',
-        }
-
-
-class ResUsers(models.Model):
-    _inherit = 'res.users'
-
-    def write(self, vals):
-        res = super().write(vals)
-        # Propagar cambios al partner y persona asociados, salvo que se indique evitar sincronización
-        if not self.env.context.get('skip_persona_sync'):
-            for user in self:
-                partner = user.partner_id
-                if not partner:
-                    continue
-                # Actualizar partner con nombre o email/login
-                upd_partner = {}
-                if 'name' in vals:
-                    upd_partner['name'] = vals.get('name') or user.name
-                # Priorizar login como email si se cambia, si no, usar email explícito
-                new_email = None
-                if 'login' in vals:
-                    new_email = vals.get('login')
-                elif 'email' in vals:
-                    new_email = vals.get('email')
-                if new_email is not None:
-                    upd_partner['email'] = new_email or False
-                if upd_partner:
-                    partner.sudo().with_context(skip_persona_sync=True).write(upd_partner)
-                # Actualizar persona vinculada: nombre desglosado y email
-                persona = self.env['erasmus.persona'].search([('partner_id', '=', partner.id)], limit=1)
-                if persona:
-                    upd_persona = {}
-                    if 'name' in vals:
-                        parts = (upd_partner.get('name') or user.name or '').split()
-                        upd_persona['nombre'] = parts[0] if parts else False
-                        upd_persona['apellido1'] = parts[-1] if len(parts) > 1 else False
-                        upd_persona['apellido2'] = ' '.join(parts[1:-1]) if len(parts) > 2 else False
-                    if new_email is not None:
-                        upd_persona['email'] = new_email or False
-                    if upd_persona:
-                        # Evitar escribir partner de nuevo y evitar reescritura sobre el propio user
-                        persona.with_context(skip_partner_sync=True, from_user=True, skip_user_sync=True).write(upd_persona)
-        return res
-
-
-class ErasmusPais(models.Model):
-    _name = 'erasmus.pais'
-    _description = 'País Erasmus (preferencias)'
-    _rec_name = 'name'
-    _order = 'name_es'
-
-    name = fields.Char(string='Nombre (EU - ES)', compute='_compute_name', store=True, index=True)
-    name_eu = fields.Char(string='Nombre es Euskera', required=True)
-    name_es = fields.Char(string='Nombre en Castellano', required=True, index=True)
-    country_id = fields.Many2one('res.country', string='País de Odoo', required=True, index=True)
-    selection_scope = fields.Selection([
-        ('ambos', 'Ambos'),
-        ('estudiante', 'Estudiante'),
-    ], string='Lo pueden seleccionar', default='ambos', required=True, index=True)
-    active = fields.Boolean(default=True)
-
-    _sql_constraints = [
-        ('uniq_country_scope', 'unique(country_id)', 'Cada país de Odoo solo puede estar una vez en la tabla de países Erasmus.'),
-    ]
-
-    @api.depends('name_eu', 'name_es')
-    def _compute_name(self):
-        for rec in self:
-            eu = rec.name_eu or ''
-            es = rec.name_es or ''
-            rec.name = f"{eu} - {es}" if eu and es else (eu or es)
-
-    def name_get(self):
-        """Usar el nombre compuesto almacenado para asegurar consistencia en cualquier parte (dropdowns, search more, etc.)."""
-        return [(rec.id, rec.name or '') for rec in self]
-
-
-class ResCountry(models.Model):
-    _inherit = 'res.country'
-
-    def name_get(self):
-        res = super().name_get()
-        if not self.env.context.get('erasmus_pais_label'):
-            return res
-        # Build map country_id -> (eu, es)
-        Pais = self.env['erasmus.pais'].sudo()
-        mapping = {p.country_id.id: (p.name_eu or '', p.name_es or '') for p in Pais.search([('country_id', 'in', self.ids)])}
-        labeled = []
-        for rec in self:
-            eu_es = mapping.get(rec.id)
-            if eu_es:
-                eu, es = eu_es
-                label = f"{eu} - {es}" if eu and es else (eu or es or rec.name)
-                labeled.append((rec.id, label))
-            else:
-                labeled.append((rec.id, rec.name))
-        return labeled
